fix: render of triangle bnow back in 2d space

+++ src/Runtime/src/Backends/OpenGL/OpenGLVertexBuffer.cpp
@@ -8,7 +8,8 @@
 namespace Neon
 {
     OpenGLVertexBuffer::OpenGLVertexBuffer(const float* vertices, const size_t size)
-        : m_vertices(vertices), m_size(size)
+        : m_vertices(vertices)
+        , m_size(size)
     {
         glGenVertexArrays(1, &m_VAO);
         glGenBuffers(1, &m_VBO);
@@ -43,12 +44,8 @@ namespace Neon
     {
         glBindVertexArray(m_VAO);
         glBindBuffer(GL_ARRAY_BUFFER, m_VBO);
-        // needs work
-        glBufferData(GL_ARRAY_BUFFER, static_cast<size_t>(m_size), m_vertices, GL_STATIC_DRAW);
-        //glBufferData(GL_ARRAY_BUFFER, m_size * sizeof(float), m_vertices, GL_STATIC_DRAW);
+        glBufferData(GL_ARRAY_BUFFER, m_size, m_vertices, GL_STATIC_DRAW);
 
-
-        // Set the vertex attribute pointer
         glVertexAttribPointer(0, 2, GL_FLOAT, GL_FALSE, 2 * sizeof(GLfloat), (GLvoid*)0);
         glEnableVertexAttribArray(0);
     }
